requirements.txt
fastmcp>=0.4.1
mcp[cli]>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
pyjwt>=2.8.0
